edge/requirements.txt
mediapipe>=0.10.9
opencv-python>=4.8.0
numpy>=1.24.0
requests>=2.31.0
python-dotenv>=1.0.0
# depthai>=2.24.0  # Uncomment on Jetson with OAK-D Lite connected
